[c++ bigint] Optimize equality/inequality operators
No need to run the comparison twice.

Also, starting from the least significant chunk results in a much faster comparison of numbers that are close to each other. It shouldn't influence the other cases, as there's little chance that two numbers with the same number of chunks will differ only in the most significant bits -- if they differ, they will most probably differ in the least significant bits as well.